Extract the packaging-verify in memory CPI,CPB,CPB builders into new library (packaging-test-utilities)
