+++ libs/packaging/packaging-verify/src/test/kotlin/net/corda/libs/packaging/verify/internal/cpi/CpiV1VerifierTest.kt
@@ -5,14 +5,14 @@ import net.corda.libs.packaging.core.exception.CordappManifestException
 import net.corda.libs.packaging.core.exception.DependencyResolutionException
 import net.corda.libs.packaging.core.exception.InvalidSignatureException
 import net.corda.test.util.InMemoryZipFile
-import net.corda.libs.packaging.verify.TestUtils
-import net.corda.libs.packaging.verify.TestUtils.ALICE
-import net.corda.libs.packaging.verify.TestUtils.BOB
-import net.corda.libs.packaging.verify.TestUtils.ROOT_CA
-import net.corda.libs.packaging.verify.TestUtils.signedBy
-import net.corda.libs.packaging.verify.internal.cpb.TestCpbV1Builder
-import net.corda.libs.packaging.verify.internal.cpb.TestCpbV1Builder.Companion.POLICY_FILE
-import net.corda.libs.packaging.verify.internal.cpk.TestCpkV1Builder
+import net.corda.libs.packaging.testutils.TestUtils
+import net.corda.libs.packaging.testutils.TestUtils.ALICE
+import net.corda.libs.packaging.testutils.TestUtils.BOB
+import net.corda.libs.packaging.testutils.TestUtils.ROOT_CA
+import net.corda.libs.packaging.testutils.TestUtils.signedBy
+import net.corda.libs.packaging.testutils.cpb.TestCpbV1Builder
+import net.corda.libs.packaging.testutils.cpb.TestCpbV1Builder.Companion.POLICY_FILE
+import net.corda.libs.packaging.testutils.cpk.TestCpkV1Builder
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
