+++ settings.gradle
@@ -352,6 +352,7 @@ include 'testing:flow:dummy-link-manager'
 include 'testing:layered-property-map-testkit'
 include 'testing:message-patterns'
 include 'testing:p2p:inmemory-messaging-impl'
+include 'testing:packaging-test-utilities'
 include 'testing:sandboxes'
 include 'testing:security-manager-utilities'
 include 'testing:test-utilities'
