+++ testing/packaging-test-utilities/src/main/kotlin/net/corda/libs/packaging/testutils/cpb/TestCpbV2Builder.kt
@@ -1,14 +1,14 @@
-package net.corda.libs.packaging.verify.internal.cpb
+package net.corda.libs.packaging.testutils.cpb
 
+import net.corda.libs.packaging.testutils.TestUtils
+import net.corda.libs.packaging.testutils.TestUtils.addFile
+import net.corda.libs.packaging.testutils.TestUtils.signedBy
+import net.corda.libs.packaging.testutils.cpk.TestCpkV2Builder
 import net.corda.test.util.InMemoryZipFile
-import net.corda.libs.packaging.verify.TestUtils
-import net.corda.libs.packaging.verify.TestUtils.addFile
-import net.corda.libs.packaging.verify.TestUtils.signedBy
-import net.corda.libs.packaging.verify.internal.cpk.TestCpkV2Builder
 import java.io.ByteArrayInputStream
 import java.util.jar.Manifest
 
-internal class TestCpbV2Builder {
+class TestCpbV2Builder {
     var name = "testCpbV2.cpb"
         private set
     var version = "1.0.0.0"
