+++ testing/packaging-test-utilities/src/main/kotlin/net/corda/libs/packaging/testutils/cpk/TestCpkV1MainBundleBuilder.kt
@@ -1,16 +1,16 @@
-package net.corda.libs.packaging.verify.internal.cpk
+package net.corda.libs.packaging.testutils.cpk
 
 import net.corda.libs.packaging.PackagingConstants.CPK_BUNDLE_NAME_ATTRIBUTE
 import net.corda.libs.packaging.PackagingConstants.CPK_BUNDLE_VERSION_ATTRIBUTE
 import net.corda.test.util.InMemoryZipFile
-import net.corda.libs.packaging.verify.TestUtils
-import net.corda.libs.packaging.verify.TestUtils.addFile
-import net.corda.libs.packaging.verify.TestUtils.signedBy
-import net.corda.libs.packaging.verify.TestUtils.toBase64
+import net.corda.libs.packaging.testutils.TestUtils
+import net.corda.libs.packaging.testutils.TestUtils.addFile
+import net.corda.libs.packaging.testutils.TestUtils.signedBy
+import net.corda.libs.packaging.testutils.TestUtils.toBase64
 import java.io.ByteArrayInputStream
 import java.util.jar.Manifest
 
-internal class TestCpkV1MainBundleBuilder {
+class TestCpkV1MainBundleBuilder {
     var name = "testMainBundle.jar"
         private set
     var bundleName = "test.cpk"
