+++ testing/packaging-test-utilities/src/main/resources/alice.p12
