+++ testing/packaging-test-utilities/src/main/resources/generate-keystores.sh
